Fix floating point detection for 64bit values
I forgot to add this file in 7d85d7a41bf31c686fba697d4c4107cf7eebbda5.

+++ src/sqlitetablemodel.cpp
@@ -486,7 +486,7 @@ bool SqliteTableModel::setTypedData(const QModelIndex& index, bool isBlob, const
                 type = SQLITE_INTEGER;
         } else if(m_vDataTypes.at(column) == SQLITE_FLOAT) {
             bool ok;
-            newValue.toFloat(&ok);
+            newValue.toDouble(&ok);
             if(ok)
                 type = SQLITE_FLOAT;
         }
